refactor: store melee constants as JSON
