+++ src/melee/animationUtils.ts
