+++ src/melee/moves.json
@@ -1,0 +1,122 @@
+{
+  "1": {
+    "name": "Miscellaneous",
+    "shortName": "misc"
+  },
+  "2": {
+    "name": "Jab",
+    "shortName": "jab"
+  },
+  "3": {
+    "name": "Jab",
+    "shortName": "jab"
+  },
+  "4": {
+    "name": "Jab",
+    "shortName": "jab"
+  },
+  "5": {
+    "name": "Rapid Jabs",
+    "shortName": "rapid-jabs"
+  },
+  "6": {
+    "name": "Dash Attack",
+    "shortName": "dash"
+  },
+  "7": {
+    "name": "Forward Tilt",
+    "shortName": "ftilt"
+  },
+  "8": {
+    "name": "Up Tilt",
+    "shortName": "utilt"
+  },
+  "9": {
+    "name": "Down Tilt",
+    "shortName": "dtilt"
+  },
+  "10": {
+    "name": "Forward Smash",
+    "shortName": "fsmash"
+  },
+  "11": {
+    "name": "Up Smash",
+    "shortName": "usmash"
+  },
+  "12": {
+    "name": "Down Smash",
+    "shortName": "dsmash"
+  },
+  "13": {
+    "name": "Neutral Air",
+    "shortName": "nair"
+  },
+  "14": {
+    "name": "Forward Air",
+    "shortName": "fair"
+  },
+  "15": {
+    "name": "Back Air",
+    "shortName": "bair"
+  },
+  "16": {
+    "name": "Up Air",
+    "shortName": "uair"
+  },
+  "17": {
+    "name": "Down Air",
+    "shortName": "dair"
+  },
+  "18": {
+    "name": "Neutral B",
+    "shortName": "neutral-b"
+  },
+  "19": {
+    "name": "Side B",
+    "shortName": "side-b"
+  },
+  "20": {
+    "name": "Up B",
+    "shortName": "up-b"
+  },
+  "21": {
+    "name": "Down B",
+    "shortName": "down-b"
+  },
+  "50": {
+    "name": "Getup Attack",
+    "shortName": "getup"
+  },
+  "51": {
+    "name": "Getup Attack (Slow)",
+    "shortName": "getup-slow"
+  },
+  "52": {
+    "name": "Grab Pummel",
+    "shortName": "pummel"
+  },
+  "53": {
+    "name": "Forward Throw",
+    "shortName": "fthrow"
+  },
+  "54": {
+    "name": "Back Throw",
+    "shortName": "bthrow"
+  },
+  "55": {
+    "name": "Up Throw",
+    "shortName": "uthrow"
+  },
+  "56": {
+    "name": "Down Throw",
+    "shortName": "dthrow"
+  },
+  "61": {
+    "name": "Edge Attack (Slow)",
+    "shortName": "edge-slow"
+  },
+  "62": {
+    "name": "Edge Attack",
+    "shortName": "edge"
+  }
+}
