+++ src/melee/stages.json
@@ -1,0 +1,59 @@
+{
+  "2": "Fountain of Dreams",
+  "3": "Pokémon Stadium",
+  "4": "Princess Peach's Castle",
+  "5": "Kongo Jungle",
+  "6": "Brinstar",
+  "7": "Corneria",
+  "8": "Yoshi's Story",
+  "9": "Onett",
+  "10": "Mute City",
+  "11": "Rainbow Cruise",
+  "12": "Jungle Japes",
+  "13": "Great Bay",
+  "14": "Hyrule Temple",
+  "15": "Brinstar Depths",
+  "16": "Yoshi's Island",
+  "17": "Green Greens",
+  "18": "Fourside",
+  "19": "Mushroom Kingdom I",
+  "20": "Mushroom Kingdom II",
+  "22": "Venom",
+  "23": "Poké Floats",
+  "24": "Big Blue",
+  "25": "Icicle Mountain",
+  "26": "Icetop",
+  "27": "Flat Zone",
+  "28": "Dream Land N64",
+  "29": "Yoshi's Island N64",
+  "30": "Kongo Jungle N64",
+  "31": "Battlefield",
+  "32": "Final Destination",
+  "33": "Target Test (Mario)",
+  "34": "Target Test (Captain Falcon)",
+  "35": "Target Test (Young Link)",
+  "36": "Target Test (Donkey Kong)",
+  "37": "Target Test (Dr. Mario)",
+  "38": "Target Test (Falco)",
+  "39": "Target Test (Fox)",
+  "40": "Target Test (Ice Climbers)",
+  "41": "Target Test (Kirby)",
+  "42": "Target Test (Bowser)",
+  "43": "Target Test (Link)",
+  "44": "Target Test (Luigi)",
+  "45": "Target Test (Marth)",
+  "46": "Target Test (Mewtwo)",
+  "47": "Target Test (Ness)",
+  "48": "Target Test (Peach)",
+  "49": "Target Test (Pichu)",
+  "50": "Target Test (Pikachu)",
+  "51": "Target Test (Jigglypuff)",
+  "52": "Target Test (Samus)",
+  "53": "Target Test (Sheik)",
+  "54": "Target Test (Yoshi)",
+  "55": "Target Test (Zelda)",
+  "56": "Target Test (Mr. Game & Watch)",
+  "57": "Target Test (Roy)",
+  "58": "Target Test (Ganondorf)",
+  "84": "Home-Run Contest"
+}
